requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.28.2
psycopg2==2.9.9
pycparser==2.22
python-decouple==3.8
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
